pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pyrogram = "^2.0.106"
requests = "^2.31.0"
termcolor = "^2.4.0"
colored = "^2.2.4"
configs = "^3.0.3"
db = "^0.1.1"
luhn = "^0.2.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"
telethon = "^1.33.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"